Segment files into modules
