+++ package.json
@@ -1,0 +1,16 @@
+{
+  "name": "minesweeper-terminal",
+  "version": "1.0.0",
+  "description": "Play the game Minesweeper through the Terminal",
+  "main": "index.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1",
+    "build": "babel src -d lib"
+  },
+  "author": "",
+  "license": "ISC",
+  "devDependencies": {
+    "babel-cli": "^6.26.0",
+    "babel-preset-env": "^1.7.0"
+  }
+}
