+++ src/game.js
@@ -1,0 +1,41 @@
+
+/*
+* To play Minesweeper, we will create instances of MineSweeperGame in command line.
+* For example:
+* In the command line, navigate to the lib directory and run `node`
+* Run `.load game.js` to load the contents of this file.
+* Then create a Game instance and run commands like so:
+* let game = new Game(3, 3, 3);
+* game.playMove(0, 1);
+* game.playMove(1, 2);
+* When done run `.exit`
+*/
+
+
+import Board from './board';
+
+class Game {
+    constructor(numberOfRows, numberOfColumns, numberOfBombs) {
+        this._board = new Board(numberOfRows, numberOfColumns, numberOfBombs);
+
+    }
+    playMove(rowIndex, columnIndex){
+
+        this._board.flipTile(rowIndex, columnIndex);
+        
+        // if bomb at flip location, player loses
+        if (this._board.playerBoard[rowIndex][columnIndex] === 'B') {
+            console.log('Game is over! Final Board:');
+            this._board.print()
+
+        // if no bombs left and player still safe, player wins
+         } else if (!this._board.hasSafeTiles()) {        
+            console.log('You won!');
+         // if no bombs and game not over, print board to continue   
+         } else {
+             console.log('Current Board: ');
+             this._board.print();
+
+         }
+    }
+}
